+++ src/onegov.org/onegov/org/views/event.py
@@ -8,7 +8,7 @@ from onegov.org import _, OrgApp
 from onegov.org.elements import Link
 from onegov.org.forms import EventForm
 from onegov.org.layout import EventLayout
-from onegov.org.mail import send_html_mail
+from onegov.org.mail import send_transactional_html_mail
 from onegov.org.models import TicketMessage, EventMessage
 from onegov.ticket import TicketCollection
 from sedate import utcnow
@@ -62,7 +62,7 @@ def publish_event(self, request):
     ticket = TicketCollection(session).by_handler_id(self.id.hex)
 
     if self.meta['submitter_email'] != request.current_username:
-        send_html_mail(
+        send_transactional_html_mail(
             request=request,
             template='mail_event_accepted.pt',
             subject=_("Your event was accepted"),
@@ -160,7 +160,7 @@ def view_event(self, request):
             TicketMessage.create(ticket, request, 'opened')
 
         if self.meta['submitter_email'] != request.current_username:
-            send_html_mail(
+            send_transactional_html_mail(
                 request=request,
                 template='mail_ticket_opened.pt',
                 subject=_("A ticket has been opened"),
@@ -232,7 +232,7 @@ def handle_delete_event(self, request):
         ticket.create_snapshot(request)
 
     if self.meta['submitter_email'] != request.current_username:
-        send_html_mail(
+        send_transactional_html_mail(
             request=request,
             template='mail_event_rejected.pt',
             subject=_("Your event was rejected"),